ml-on-the-fly/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
python-dotenv
python-multipart
requests
pydantic 